+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.1.0] - 2021-05-28
+
+Add `useUrlState` to change both *Route Params* and *Query String*.
+
+### Added
+
+- `useUrlState`: with type `(defaultValues?: UrlState): [UrlState, DisaptchState<UrlState>]`.
+
 ## [4.0.0] - 2021-05-26
 
 Update `useQueryStringKey` to allow dispatch.